requirements.txt
aiogram>=3.0
asyncpg
pyyaml
python-dotenv
aiohttp
aiofiles
